Moved `Target` into its own module
This is a preliminary step to make `Target` a first-class enum that
can be passed to public functions.

This can be used to remove duplication as most of the logic to get
intensities for a region or a postcode is the same (only API URL really
differs).
